Fix framerate
http://www.raspberrypi.org/new-camera-mode-released/
One thing to note: the system will always try to run at the frame rate
specified in preference to resolution. Therefore if you specify a high
rate at a resolution it cannot manage, it will use a low resolution mode
to achieve the frame rate and upscale to the requested size – upscaling
rarely looks good. It may also be too fast for the video encoder, so
some of the extra frames may be skipped. So always ensure the resolution
you specify can achieve the required frame rate to get the best results.